+++ app/src/main/provider-settings/header.txt
@@ -1,0 +1,3 @@
+/*
+ * This source file is generated with https://github.com/BoD/android-contentprovider-generator
+ */
