+++ app/src/main/provider-settings/item.json
@@ -1,0 +1,20 @@
+{
+    "fields": [
+	{
+	    "name": "title",
+	    "type": "String",
+	    "nullable": false,
+	    "index": true,
+	},
+	{
+	    "name": "description",
+	    "type": "String",
+	    "nullable": false,
+	},
+	{
+	    "name": "creationDate",
+	    "type": "Integer",
+	    "nullable": false
+	}
+    ]
+}
